Added Order Service

+++ src/main/java/com/rgiftings/Backend/Service/ProductService.java
@@ -23,7 +23,7 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public void deleteProduct(int id) {
+    public void deleteProduct(Long id) {
         productRepository.deleteById(id);
     }
 }
